refactor: 불필요한 주석 제거, 코드 정리
