+++ src/app/(with-nav-footer)/activities/[id]/ActivityDetail/_components/ActivityReservation/ActivityReservationStyles.css
@@ -1,5 +1,4 @@
 .react-calendar {
-  /* border:none; */
   border-color:#ddd;
   padding: 5px 10px;
   border-radius: 12px;
@@ -36,7 +35,6 @@
   }
 }
 
-
 .react-calendar__navigation {
   margin-bottom: 0!important;
 }
@@ -67,7 +65,6 @@
   color:#4b4b4b;
 }
 
-/* 오늘 날짜 스타일 */
 .react-calendar__tile--now {
   background-color: #0b3b2d;
   color: #ffffff!important;
@@ -75,10 +72,9 @@
   text-decoration: inherit!important;
 }
 
-/* 오늘 날짜에 밑줄만 추가 */
 .react-calendar__tile--today-underline {
-  color: #0b3b2d!important; /* 기본 글자 색 */
-  background-color: #ced8d5!important; /* 예약 가능한 날짜 배경 색 */
+  color: #0b3b2d!important; 
+  background-color: #ced8d5!important; 
   border-radius: 8px;
 }
 
@@ -97,23 +93,20 @@
   background:none;
 }
 
+.react-calendar__tile--available:hover {
+  background-color: #0b3b2d!important; 
+  color: #fff!important;
+}
 
-  
-  /* 예약 가능한 날짜 호버 스타일 */
-  .react-calendar__tile--available:hover {
-    background-color: #0b3b2d!important; 
-    color: #fff!important;
-  }
+.react-calendar__month-view__days__day--neighboringMonth, .react-calendar__decade-view__years__year--neighboringDecade, .react-calendar__century-view__decades__decade--neighboringCentury {
+  color:#a4a1a4!important;
+}
+.react-calendar__month-view__days__day--weekend{
+  color:#4b4b4b;
+}
 
-  .react-calendar__month-view__days__day--neighboringMonth, .react-calendar__decade-view__years__year--neighboringDecade, .react-calendar__century-view__decades__decade--neighboringCentury {
-    color:#a4a1a4!important;
-  }
-  .react-calendar__month-view__days__day--weekend{
-    color:#4b4b4b;
-  }
-/* 예약 가능한 날짜 스타일 */
 .react-calendar__tile--available {
-  color: #0b3b2d; /* 예약 가능한 날짜 글자 색 */
+  color: #0b3b2d; 
   border-radius: 8px;
   text-decoration: underline;
   font-weight: bold;
@@ -125,10 +118,10 @@
 
 @media (max-width: 860px) {
   .custom-md-size-top span, .custom-md-size-middle p:first-child, .custom-md-size-bottom span:first-child {
-    font-size: 16px; /* 860px 이하에서 크기 줄이기 */
+    font-size: 16px; 
   }
   .custom-md-size-top, .custom-md-size-bottom span:last-child {
-    font-size: 24px; /* 860px 이하에서 크기 줄이기 */
+    font-size: 24px;
   }
   .custom-md-size-middle .second {
     font-size: 14px;
